requirements.txt
Collectfast==0.1.13
Django==1.5
boto==2.20.1
dj-database-url==0.3.0
dj-static==0.0.5
django-storages==1.1.8
django-tastypie==0.10.0
django-tastypie-swagger==0.1.1
mimeparse==0.1.3
psycopg2==2.5.3
pystache==0.5.4
python-dateutil==2.2
pytz==2014.3
redis==2.9.1
rq==0.4.6
six==1.6.1
static==1.0.2
wsgiref==0.1.2
